Adress review comments
